Create a controller to artworks
