+++ config/routes.php
@@ -3,6 +3,7 @@
 use App\Controllers\AuthController;
 use App\Controllers\ClientController;
 use App\Controllers\ArtistController;
+use App\Controllers\ArtworkController;
 use App\Controllers\HomeController;
 use Core\Router\Route;
 
@@ -22,10 +23,10 @@ Route::middleware('auth')->group(function () {
     });
 
     Route::middleware('artist')->group(function () {
-        Route::get('/createPost', [ArtistController::class, 'new'])->name('artist.new');
+        Route::get('/createPost', [ArtworkController::class, 'new'])->name('artist.new');
     });
 
     Route::middleware('artist')->group(function () {
-        Route::post('/upload', [ArtistController::class, 'newArtwork'])->name('artist.newArtwork');
+        Route::post('/upload', [ArtworkController::class, 'newArtwork'])->name('artist.newArtwork');
     });
 });
